ia_aupa/requirements.txt
streamlit
requests
python-dotenv
git+https://github.com/ayrshare/social-post-api-python.git